refactor(profile): inline notifications + idea archive on desktop; strip UserMenu dropdown
The redesigned /profile already covers every action the TopNav
dropdown used to host (Profile entry + Sign out + Delete), so the
dropdown was just an extra tap. UserMenu is now a plain Avatar link
to /profile.

Desktop sidebar tabs no longer navigate away — instead they swap the
main panel content in place:
- Profile tab     → avatar hero + picker + profile card (as before)
- Notifications   → renders <NotificationsPanel /> inline
- Idea archive    → renders <ArchivedIdeasPanel /> inline

Mobile is unchanged: the sidebar is hidden, so the Preferences card
still navigates to the dedicated /profile/notifications and
/profile/archived-ideas pages.

The page bodies were extracted into shared panel components
(src/components/profile/) so the mobile dedicated pages and the
desktop inline rendering use the same source of truth.

Also tightened the profile-section responsive class — the old
"hidden md:block" branch was inverted; now uses "md:hidden" when
non-profile tab is active so desktop hides it properly without
hiding it on mobile.
